+++ src/drivers/net/natsemi.c
@@ -536,7 +536,8 @@ natsemi_init_rxd(struct nic *nic __unused)
         rxd[i].bufptr = virt_to_bus(&rxb[i*RX_BUF_SIZE]);
         if (natsemi_debug > 1)
             printf("natsemi_init_rxd: rxd[%d]=%p link=%X cmdsts=%X bufptr=%4.4x\n", 
-                   i, &rxd[i], rxd[i].link, rxd[i].cmdsts, rxd[i].bufptr);
+                   i, &rxd[i], (unsigned int) rxd[i].link, (unsigned int) rxd[i].cmdsts, 
+		   (unsigned int) rxd[i].bufptr);
     }
 
     /* load Receive Descriptor Register */
@@ -648,11 +649,11 @@ natsemi_transmit(struct nic  *nic,
         /* wait */ ;
 
     if (currticks() >= to) {
-        printf("natsemi_transmit: TX Timeout! Tx status %X.\n", tx_status);
+        printf("natsemi_transmit: TX Timeout! Tx status %X.\n", (unsigned int) tx_status);
     }
 
     if (!(tx_status & 0x08000000)) {
-	printf("natsemi_transmit: Transmit error, Tx status %X.\n", tx_status);
+	printf("natsemi_transmit: Transmit error, Tx status %X.\n", (unsigned int) tx_status);
     }
 }
 
@@ -677,7 +678,7 @@ natsemi_poll(struct nic *nic, int retrieve)
     int retstat = 0;
 
     if (natsemi_debug > 2)
-        printf("natsemi_poll: cur_rx:%d, status:%X\n", cur_rx, rx_status);
+        printf("natsemi_poll: cur_rx:%d, status:%X\n", cur_rx, (unsigned int) rx_status);
 
     if (!(rx_status & OWN))
         return retstat;
@@ -686,14 +687,14 @@ natsemi_poll(struct nic *nic, int retrieve)
 
     if (natsemi_debug > 1)
         printf("natsemi_poll: got a packet: cur_rx:%d, status:%X\n",
-               cur_rx, rx_status);
+               cur_rx, (unsigned int) rx_status);
 
     nic->packetlen = (rx_status & DSIZE) - CRC_SIZE;
 
     if ((rx_status & (DescMore|DescPktOK|RxTooLong)) != DescPktOK) {
         /* corrupted packet received */
         printf("natsemi_poll: Corrupted packet received, buffer status = %X\n",
-               rx_status);
+               (unsigned int) rx_status);
         retstat = 0;
     } else {
         /* give packet to higher level routine */
